Pulled out files into their own class
Pulled out property functions into their own class
Pulled out cursed reading of ZMod's file into it's own class
